requirements.txt
colorama==0.4.6
iniconfig==2.0.0
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
pytest==7.3.2
